+++ Steps/Serialization/Json/src/SerializeObjectToStreamStep{TObject}.cs
@@ -26,6 +26,7 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json
     using Microsoft.Extensions.Logging;
     using MyTrout.Pipelines;
     using MyTrout.Pipelines.Core;
+    using System;
     using System.Collections.Generic;
     using System.IO;
     using System.Text.Json;
@@ -36,6 +37,7 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json
     /// </summary>
     /// <typeparam name="TObject">The target object for deserialization.
     /// </typeparam>
+    [Obsolete("Use the MyTrout.Pipelines.Steps.Serialization.Core library containing SerializeObjectToStreamStep<TObject> and the SerializeJsonObjectToStreamOptions from this library.")]
     public class SerializeObjectToStreamStep<TObject> : AbstractCachingPipelineStep<SerializeObjectToStreamStep<TObject>, SerializeObjectToStreamOptions>
         where TObject : class
     {