+++ Steps/Serialization/Json/test/DeserializeJsonObjectFromStreamOptionsTests.cs
@@ -1,4 +1,4 @@
-// <copyright file="DeserializeObjectFromStreamStepTests.cs" company="Chris Trout">
+﻿// <copyright file="DeserializeJsonObjectFromStreamOptionsTests.cs" company="Chris Trout">
 // MIT License
 //
 // Copyright(c) 2022 Chris Trout
@@ -28,58 +28,28 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json.Tests
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Moq;
     using MyTrout.Pipelines.Core;
-    using System.Globalization;
     using System.IO;
     using System.Linq;
     using System.Text;
     using System.Threading.Tasks;
+    using CORE = MyTrout.Pipelines.Steps.Serialization;
 
-#pragma warning disable CS0618
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
     [TestClass]
-    public class DeserializeObjectFromStreamStepTests
+    public class DeserializeJsonObjectFromStreamOptionsTests
     {
         [TestMethod]
-        public void Constructs_DeserializeObjectFromStreamStep_Successfully()
+        public void Constructs_DeserializeJsonObjectFromStreamOptions_Successfully()
         {
             // arrange
-            var logger = new Mock<ILogger<DeserializeObjectFromStreamStep<object>>>().Object;
-            var options = new DeserializeObjectFromStreamOptions();
-            var next = new Mock<IPipelineRequest>().Object;
+            var options = new System.Text.Json.JsonSerializerOptions();
 
             // act
-            var result = new DeserializeObjectFromStreamStep<object>(logger, options, next);
+            var sut = new DeserializeJsonObjectFromStreamOptions() { JsonSerializerOptions = options };
 
             // assert
-            Assert.IsNotNull(result);
-            Assert.AreEqual(logger, result.Logger);
-            Assert.AreEqual(options, result.Options);
-            Assert.AreEqual(next, result.Next);
-        }
-
-        [TestMethod]
-        public async Task Returns_PipelineContext_Error_From_InvokeAsync_When_InputStream_Is_Not_In_Context()
-        {
-            // arrange
-            var logger = new Mock<ILogger<DeserializeObjectFromStreamStep<object>>>().Object;
-            var options = new DeserializeObjectFromStreamOptions();
-            var next = new Mock<IPipelineRequest>().Object;
-
-            var context = new PipelineContext();
-
-            var expectedErrorCount = 1;
-
-            var expectedMessage = Resources.NO_KEY_IN_CONTEXT(CultureInfo.CurrentCulture, PipelineContextConstants.INPUT_STREAM);
-
-            using (var sut = new DeserializeObjectFromStreamStep<object>(logger, options, next))
-            {
-                // act
-                await sut.InvokeAsync(context);
-            }
-
-            // assert
-            Assert.AreEqual(expectedErrorCount, context.Errors.Count);
-            Assert.AreEqual(expectedMessage, context.Errors[0].Message);
+            Assert.IsNotNull(sut);
+            Assert.AreEqual(options, sut.JsonSerializerOptions);
         }
 
         [TestMethod]
@@ -89,8 +59,8 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json.Tests
             var context = new PipelineContext();
             var expectedResult = new SampleItem() { ConnectionString = "value" };
 
-            var logger = new Mock<ILogger<DeserializeObjectFromStreamStep<SampleItem>>>().Object;
-            var options = new DeserializeObjectFromStreamOptions();
+            var logger = new Mock<ILogger<CORE.DeserializeObjectFromStreamStep<SampleItem>>>().Object;
+            var options = new DeserializeJsonObjectFromStreamOptions();
             var nextMock = new Mock<IPipelineRequest>();
             nextMock.Setup(x => x.InvokeAsync(context))
                                        .Callback(() =>
@@ -115,7 +85,7 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json.Tests
 
             var expectedErrorCount = 0;
 
-            using (var sut = new DeserializeObjectFromStreamStep<SampleItem>(logger, options, next))
+            using (var sut = new CORE.DeserializeObjectFromStreamStep<SampleItem>(logger, options, next))
             {
                 // act
                 await sut.InvokeAsync(context);
@@ -137,8 +107,8 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json.Tests
             var context = new PipelineContext();
             var expectedResult = new SampleItem() { ConnectionString = "value" };
 
-            var logger = new Mock<ILogger<DeserializeObjectFromStreamStep<SampleItem>>>().Object;
-            var options = new DeserializeObjectFromStreamOptions()
+            var logger = new Mock<ILogger<CORE.DeserializeObjectFromStreamStep<SampleItem>>>().Object;
+            var options = new DeserializeJsonObjectFromStreamOptions()
             {
                 InputStreamContextName = "TESTING_OTHER_INPUT_NAME",
                 OutputObjectContextName = "TESTING_OTHER_OUTPUT_STREAM"
@@ -167,7 +137,7 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json.Tests
 
             var expectedErrorCount = 0;
 
-            using (var sut = new DeserializeObjectFromStreamStep<SampleItem>(logger, options, next))
+            using (var sut = new CORE.DeserializeObjectFromStreamStep<SampleItem>(logger, options, next))
             {
                 // act
                 await sut.InvokeAsync(context);
@@ -186,8 +156,8 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json.Tests
         public async Task Returns_PipelineContext_From_InvokeAsync_When_Stream_Is_Deserialized_Successfully()
         {
             // arrange
-            var logger = new Mock<ILogger<DeserializeObjectFromStreamStep<SampleItem>>>().Object;
-            var options = new DeserializeObjectFromStreamOptions();
+            var logger = new Mock<ILogger<CORE.DeserializeObjectFromStreamStep<SampleItem>>>().Object;
+            var options = new DeserializeJsonObjectFromStreamOptions();
             var next = new Mock<IPipelineRequest>().Object;
             string contents = "{ \"ConnectionString\": \"value\" }";
 
@@ -198,7 +168,7 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json.Tests
 
             var expectedErrorCount = 0;
 
-            using (var sut = new DeserializeObjectFromStreamStep<SampleItem>(logger, options, next))
+            using (var sut = new CORE.DeserializeObjectFromStreamStep<SampleItem>(logger, options, next))
             {
                 // act
                 await sut.InvokeAsync(context);
@@ -212,38 +182,5 @@ namespace MyTrout.Pipelines.Steps.Serialization.Json.Tests
 
             Assert.AreEqual(expectedErrorCount, context.Errors.Count);
         }
-
-        [TestMethod]
-        public async Task Returns_PipelineContext_From_InvokeAsync_With_Previously_Configured_Values_Restored()
-        {
-            // arrange
-            var logger = new Mock<ILogger<DeserializeObjectFromStreamStep<SampleItem>>>().Object;
-            var options = new DeserializeObjectFromStreamOptions();
-            var next = new Mock<IPipelineRequest>().Object;
-            string contents = "{ \"ConnectionString\": \"value\" }";
-
-            var expectedInputStream = new MemoryStream(Encoding.UTF8.GetBytes(contents));
-            var expectedOutputObject = new object();
-
-            var context = new PipelineContext();
-            context.Items.Add(options.InputStreamContextName, expectedInputStream);
-            context.Items.Add(options.OutputObjectContextName, expectedOutputObject);
-
-            var expectedItemCount = context.Items.Count;
-
-            using (var sut = new DeserializeObjectFromStreamStep<SampleItem>(logger, options, next))
-            {
-                // act
-                await sut.InvokeAsync(context);
-            }
-
-            // assert
-            Assert.AreEqual(expectedItemCount, context.Items.Count);
-            Assert.IsTrue(context.Items.ContainsKey(options.InputStreamContextName));
-            Assert.AreEqual(expectedInputStream, context.Items[options.InputStreamContextName]);
-            Assert.IsTrue(context.Items.ContainsKey(options.OutputObjectContextName));
-            Assert.AreEqual(expectedOutputObject, context.Items[options.OutputObjectContextName]);
-        }
     }
-#pragma warning restore CS0618
 }