+++ Steps/Serialization/Json/changelog.md
@@ -1,4 +1,8 @@
 # MyTrout.Pipelines.Steps.Serialization.Json Change Log
 
+## 1.1.0
+- Mark all existing Steps and Options with the Obsolete attribute to encourage usage of the new functionality.
+- Add DeserializeJsonObjectFromStreamOptions and SerializeJsonObjectToStreamOptions for use with functionality in Pipelines.Steps.Serialization.Core library.
+
 ## 1.0.0
 - Initial implementation of DeserializeObjectFromStreamStep and SerializeObjectToStreamStep.